step-02: rename original test spec, init test
